- design - price is now float instead of int - added new PostItView - configured PostItView to be used as table footer of the menu table if restaurant notes are set - reconfigured creation of info string - configured background of menu cells (closes #314) - reconfigured the overview menu string (closes #372)
git-svn-id: http://svn.rtfnt.com/svn/studifutter/dev/trunk/app@111 dd375700-57de-49ca-a908-404e403163fd

+++ Studifutter/Machine/_Menu.h
@@ -69,9 +69,9 @@
 @property (nonatomic, retain) NSNumber *price;
 
 
-@property short priceValue;
-- (short)priceValue;
-- (void)setPriceValue:(short)value_;
+@property float priceValue;
+- (float)priceValue;
+- (void)setPriceValue:(float)value_;
 
 //- (BOOL)validatePrice:(id*)value_ error:(NSError**)error_;
 
@@ -81,9 +81,9 @@
 @property (nonatomic, retain) NSNumber *reducedPrice;
 
 
-@property short reducedPriceValue;
-- (short)reducedPriceValue;
-- (void)setReducedPriceValue:(short)value_;
+@property float reducedPriceValue;
+- (float)reducedPriceValue;
+- (void)setReducedPriceValue:(float)value_;
 
 //- (BOOL)validateReducedPrice:(id*)value_ error:(NSError**)error_;
 
@@ -140,8 +140,8 @@
 - (NSNumber*)primitivePrice;
 - (void)setPrimitivePrice:(NSNumber*)value;
 
-- (short)primitivePriceValue;
-- (void)setPrimitivePriceValue:(short)value_;
+- (float)primitivePriceValue;
+- (void)setPrimitivePriceValue:(float)value_;
 
 
 
@@ -149,8 +149,8 @@
 - (NSNumber*)primitiveReducedPrice;
 - (void)setPrimitiveReducedPrice:(NSNumber*)value;
 
-- (short)primitiveReducedPriceValue;
-- (void)setPrimitiveReducedPriceValue:(short)value_;
+- (float)primitiveReducedPriceValue;
+- (void)setPrimitiveReducedPriceValue:(float)value_;
